Fix for error checking reversion in rebase
